requirements.txt
pytorch
scipy
tqdm